implemented network connectivity state
