+++ provider/src/main/java/com/raygun/raygun4android/network/ConnectivityUtils.kt
@@ -8,7 +8,9 @@ import android.net.LinkProperties
 import android.net.Network
 import android.net.NetworkCapabilities
 import android.net.NetworkInfo
+import android.telephony.TelephonyManager
 import androidx.annotation.RequiresApi
+import timber.log.Timber
 
 object ConnectivityUtils {
     fun isNetworkAvailable(context: Context): Boolean {
@@ -16,6 +18,17 @@ object ConnectivityUtils {
             isNetworkAvailable23(context)
         } else {
             isNetworkAvailableLegacy(context)
+        }
+    }
+
+    fun networkConnectivityState(context: Context): String {
+        if (!isNetworkAvailable(context)) {
+            return "Not connected"
+        }
+        return if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
+            readNetworkConnectivityState(context)
+        } else {
+            readNetworkConnectivityStateLegacy(context)
         }
     }
 
@@ -55,5 +68,64 @@ object ConnectivityUtils {
         val connectivityManager = connectivityManager(context)
         val currentNetwork = currentNetwork(context)
         return connectivityManager.getLinkProperties(currentNetwork)
+    }
+
+    @RequiresApi(android.os.Build.VERSION_CODES.M)
+    private fun readNetworkConnectivityState(context: Context): String {
+        var result = "Connected - "
+        networkCapabilities(context)?.let { capabilities ->
+            when {
+                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> result += "WiFi"
+                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> {
+                    result += "Mobile - "
+                    val linkProperties = linkProperties(context)
+                    if (linkProperties != null) {
+                        result += linkProperties.interfaceName
+                    }
+                }
+                else -> result += "unknown type"
+            }
+        }
+        Timber.d("Raygun4", "Network capabilities: $result")
+        return result
+    }
+
+    private fun readNetworkConnectivityStateLegacy(context: Context): String {
+        var result = "Connected - "
+        currentNetworkInfoLegacy(context)?.let { info ->
+            when (info.type) {
+                ConnectivityManager.TYPE_WIFI -> result += "WiFi"
+                ConnectivityManager.TYPE_WIMAX -> result += "WiMax"
+                ConnectivityManager.TYPE_MOBILE,
+                ConnectivityManager.TYPE_MOBILE_DUN,
+                ConnectivityManager.TYPE_MOBILE_HIPRI,
+                ConnectivityManager.TYPE_MOBILE_MMS,
+                ConnectivityManager.TYPE_MOBILE_SUPL,
+                    -> {
+                    result += "Mobile - "
+                    result +=
+                        when (info.subtype) {
+                            TelephonyManager.NETWORK_TYPE_1xRTT -> "1xRTT"
+                            TelephonyManager.NETWORK_TYPE_CDMA -> "CDMA"
+                            TelephonyManager.NETWORK_TYPE_EDGE -> "EDGE"
+                            TelephonyManager.NETWORK_TYPE_EVDO_0 -> "EVDO_0"
+                            TelephonyManager.NETWORK_TYPE_EVDO_A -> "EVDO_A"
+                            TelephonyManager.NETWORK_TYPE_GPRS -> "GPRS"
+                            TelephonyManager.NETWORK_TYPE_HSDPA -> "HSDPA"
+                            TelephonyManager.NETWORK_TYPE_HSPA -> "HSPA"
+                            TelephonyManager.NETWORK_TYPE_HSUPA -> "HSUPA"
+                            TelephonyManager.NETWORK_TYPE_UMTS -> "UMTS"
+                            TelephonyManager.NETWORK_TYPE_IDEN -> "IDEN"
+                            TelephonyManager.NETWORK_TYPE_UNKNOWN ->
+                                "subtype unknown/EVDO_B/EHRPD/LTE/HSPAP or similar"
+
+                            else -> "subtype unknown/EVDO_B/EHRPD/LTE/HSPAP or similar"
+                        }
+                }
+
+                else -> result += "unknown type"
+            }
+        }
+        return result
     }
 }
